refactor[#35]: global 패키지 이동 및 dto 추가
