+++ src/main/java/org/highfive/backend/global/client/fastapi/dto/RecommendContentsResponseDto.java
@@ -1,0 +1,9 @@
+package org.highfive.backend.global.client.fastapi.dto;
+
+public record RecommendContentsResponseDto(
+        long id,
+        String title,
+        String thumbnailUrl,
+        double similarity
+) {
+}
